synthia-foundry/backend/requirements.txt
fastapi
uvicorn
llama-cpp-python
